- Added ssl verify option. Now its possible to disable ssl verification. - Added Tag Split to key value - Changed Tag handling after netbox update
We updated out netbox instance. After that the following error occured:
could not be converted to string

The problem is, that they changed the api response of a tag:

[dns_name] =>
[description] => azure_reserved
[tags] => Array
	(
		[0] => tag1
		[1] => tag2
	)

[custom_fields] => stdClass Object
	(
	)

[dns_name] =>
[description] => azure_reserved
[tags] => Array
	(
		[0] => stdClass Object
			(
				[id] => 67
				[url] => https://10.138.1.131/api/extras/tags/67/
				[name] => azure_reserved
				[slug] => azure_reserved
				[color] => 9e9e9e
			)

	)

[custom_fields] => stdClass Object
	(
	)

I'm no php programmer (to be honest, I'm not a programmer at all), so my code is really bad. Maybe one of the maintainer is able to
translate my code (my bad code -> really good code)

Thanks alot
Cheers Mordecaine